placed the main header background in its own component
